+++ macros/streams/snowflake__stream.sql
@@ -10,14 +10,14 @@
   {%- set source_schema = config.get('source_schema', default=schema) -%}
   {%- set source_database = config.get('source_database', default=database) -%}
   {%- set source_type = config.get('source_type', default='internal') -%}
-  
+
   {% set target_relation = this %}
   {% set source_relation = adapter.get_relation(identifier=source_model, schema=source_schema, database=source_database) %}
 
   {% if source_relation == none %}
     {% set source_relation = api.Relation.create(identifier=source_model, schema=source_schema, database=source_database) %} 
   {% endif %}
-  
+
   -- setup
   {{ run_hooks(pre_hooks, inside_transaction=False) }}